Firebase, Menu and More
Have sorted the firebase, so far it is linked to the play button and
text box, it starts a game if no game is open else it will join a game,
it logs each player as name and opponent, it also create the initial
array in the call to the database for initialisation then applies both
of these at the update to the database push. there are some little bits
to work out but provides the functionality for the game, just to put in
a check on the text box and error message if not a proper email
structure so will get done asap. have also created the board in tiled
and is visible alongside the resign button.