board array, mouse click added and tidy
Have placed the initial structure of the game into the constants file,
boardstruc, p1 and p2 ChipSet. Have also adjusted the chip location to
point to a location within the board array rather than haveing a cell
and row for each chip. Have also added in the first of the mouse down
events, so far allows the player to select only their chip to move.