Movement and more
Have added movement into the game and updated the firebase on the move
made. There is currently no rules added to the game. I have altered some
variables and added in new constants. Firebase also has 2 added
properties turn and winner, to indicate the winner of the game and and
who is currently actively playing.