Rules, Queens, and checks
Have added all the necessary rules for the game have left out rule 6 as
it conflicted with rule 12. have added in the queen feature and the
rules governing the queen. have also added code to check if user already
exists if not play game else error message same with no name being added
error message. have added current players name to the screen so players
know who's turn it is. have tided up a lot of different functions and
deleted so code that was just a waist of space i.e. if statements.